modified file structure of frontend to fit with current implementation
