+++ apps/ca-portal/server/server/constants/location.js
@@ -1,7 +1,7 @@
 const locationMap = {
   "In-Person": "INPERSON",
   "Hybrid": "HYBRID",
-  "Online": "REMOTE",
+  "Remote": "REMOTE",
 };
 
 module.exports = {
