+++ apps/ca-portal/server/server/database/test_data/06_jobPositions.sql
@@ -1,10 +1,10 @@
 INSERT INTO `JobPosition` (`id`, `courseCode`, `sectionNumber`, `semesterCode`, `facultyUID`, `maxCAs`, `jobPositionStatus`, `location`, `locationType`, `gradeRequirement`, `graduateStatusRequirement`, `courseTakenRequirement`, `startDate`, `endDate`) VALUES
-('2241-SWEN-344-1', 'SWEN-344', 1, '2241', 200, 1, 'OPEN', 'ONLINE', 'REMOTE', 'A', Null, 1, '2025-08-25', '2025-12-25'),
+('2241-SWEN-344-1', 'SWEN-344', 1, '2241', 200, 1, 'OPEN', 'ONLINE ASYNC', 'REMOTE', 'A', Null, 1, '2025-08-25', '2025-12-25'),
 ('2241-SWEN-344-2', 'SWEN-344', 2, '2241', 250, 2, 'OPEN', 'GOL-1500', 'INPERSON', 'B_PLUS', Null, 1, '2025-08-25', '2025-12-25'),
 ('2241-SWEN-344-3', 'SWEN-344', 3, '2241', 200, 1, 'ONHOLD', 'GOL-1520', 'INPERSON', 'A_MINUS', Null, 1, '2025-08-25', '2025-12-25'),
-('2241-MATH-123-1', 'MATH-123', 1, '2241', 300, 1, 'OPEN', 'GOS-2400/ONLINE', 'HYBRID', 'A', 'UNDERGRADUATE', 1, '2025-08-25', '2025-12-25'),
+('2241-MATH-123-1', 'MATH-123', 1, '2241', 300, 1, 'OPEN', 'GOS-2400/ONLINE SYNC', 'HYBRID', 'A', 'UNDERGRADUATE', 1, '2025-08-25', '2025-12-25'),
 ('2241-SWEN-261-1', 'SWEN-261', 1, '2241', 200, 1, 'OPEN', 'GOL-1620', 'INPERSON', Null, Null, 0, '2025-08-25', '2025-12-25'),
 ('2241-SWEN-261-2', 'SWEN-261', 2, '2241', 250, 1, 'INACTIVE', 'GOL-1620', 'INPERSON', Null, Null, 0, '2025-08-25', '2025-12-25'),
 ('2241-SWEN-561-1', 'SWEN-561', 1, '2241', 250, 2, 'OPEN', 'GOL-1620', 'INPERSON', Null, 'GRADUATE', 1, '2025-08-25', '2025-12-25'),
-('2241-SWEN-561-2', 'SWEN-561', 2, '2241', 250, 1, 'FILLED', 'GOL-1620/ONLINE', 'HYBRID', Null, Null, 0, '2025-08-25', '2025-12-25'),
+('2241-SWEN-561-2', 'SWEN-561', 2, '2241', 250, 1, 'FILLED', 'GOL-1620/ONLINE SYNC', 'HYBRID', Null, Null, 0, '2025-08-25', '2025-12-25'),
 ('2241-SWEN-383-1', 'SWEN-383', 1, '2241', 200, 1, 'OPEN', 'GOL-1620', 'INPERSON', 'C_MINUS', Null, 0, '2025-08-25', '2025-12-25');
