+++ apps/ca-portal/ui/src/assets/icons.js
@@ -1,0 +1,116 @@
+// --- Reusable Icon Components ---
+export const ClockIcon = () => (
+  <svg
+    xmlns='http://www.w3.org/2000/svg'
+    className='h-5 w-5 mr-1.5 text-gray-500 inline'
+    fill='none'
+    viewBox='0 0 24 24'
+    stroke='currentColor'
+  >
+    <path
+      strokeLinecap='round'
+      strokeLinejoin='round'
+      strokeWidth={2}
+      d='M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z'
+    />
+  </svg>
+);
+export const LocationIcon = () => (
+  <svg
+    xmlns='http://www.w3.org/2000/svg'
+    className='h-5 w-5 mr-1.5 text-gray-500 inline'
+    fill='none'
+    viewBox='0 0 24 24'
+    stroke='currentColor'
+  >
+    <path
+      strokeLinecap='round'
+      strokeLinejoin='round'
+      strokeWidth={2}
+      d='M17.657 16.657L13.414 20.9a1.998 1.998 0 01-2.827 0l-4.244-4.243a8 8 0 1111.314 0z'
+    />
+    <path
+      strokeLinecap='round'
+      strokeLinejoin='round'
+      strokeWidth={2}
+      d='M15 11a3 3 0 11-6 0 3 3 0 016 0z'
+    />
+  </svg>
+);
+export const CalendarIcon = () => (
+  <svg
+    xmlns='http://www.w3.org/2000/svg'
+    className='h-5 w-5 mr-1.5 text-gray-500 inline'
+    fill='none'
+    viewBox='0 0 24 24'
+    stroke='currentColor'
+  >
+    <path
+      strokeLinecap='round'
+      strokeLinejoin='round'
+      strokeWidth={2}
+      d='M8 7V3m8 4V3m-9 8h10M5 21h14a2 2 0 002-2V7a2 2 0 00-2-2H5a2 2 0 00-2 2v12a2 2 0 002 2z'
+    />
+  </svg>
+);
+export const ResumeIcon = () => (
+  <svg
+    xmlns='http://www.w3.org/2000/svg'
+    className='h-6 w-6 text-gray-400'
+    fill='none'
+    viewBox='0 0 24 24'
+    stroke='currentColor'
+  >
+    <path
+      strokeLinecap='round'
+      strokeLinejoin='round'
+      strokeWidth='2'
+      d='M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z'
+    />
+  </svg>
+);
+
+export const CheckIcon = () => (
+  <svg
+    xmlns='http://www.w3.org/2000/svg'
+    className='h-5 w-5 text-green-500 flex-shrink-0'
+    viewBox='0 0 20 20'
+    fill='currentColor'
+  >
+    <path
+      fillRule='evenodd'
+      d='M10 18a8 8 0 100-16 8 8 0 000 16zm3.707-9.293a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z'
+      clipRule='evenodd'
+    />
+  </svg>
+);
+
+export const XIcon = () => (
+  <svg
+    xmlns='http://www.w3.org/2000/svg'
+    className='h-5 w-5 text-red-500 flex-shrink-0'
+    viewBox='0 0 20 20'
+    fill='currentColor'
+  >
+    <path
+      fillRule='evenodd'
+      d='M10 18a8 8 0 100-16 8 8 0 000 16zM8.707 7.293a1 1 0 00-1.414 1.414L8.586 10l-1.293 1.293a1 1 0 101.414 1.414L10 11.414l1.293 1.293a1 1 0 001.414-1.414L11.414 10l1.293-1.293a1 1 0 00-1.414-1.414L10 8.586 8.707 7.293z'
+      clipRule='evenodd'
+    />
+  </svg>
+);
+
+export const FilterIcon = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    className="h-5 w-5 mr-2"
+    viewBox="0 0 20 20"
+    fill="currentColor"
+  >
+    <path
+      fillRule="evenodd"
+      d="M3 3a1 1 0 011-1h12a1 1 0 011 1v3a1 1 0 01-.293.707L12 11.414V15a1 1 0 01-.293.707l-2 2A1 1 0 018 17v-5.586L3.293 6.707A1 1 0 013 6V3z"
+      clipRule="evenodd"
+    />
+  </svg>
+);
