+++ apps/ca-portal/ui/src/components/jobs/PositionsCard.js
@@ -2,40 +2,10 @@
 
 import { useAuth } from '@/contexts/AuthContext';
 import Tooltip from '../ui/ToolTip';
-import ApplicationForm from './JobPositionApplicationForm';
+import EditableApplicationForm from '../jobs/EditableApplicationForm';
 import { useState, useMemo } from 'react';
-
-// --- Icon components for the position card ---
-const CheckIcon = () => (
-  <svg xmlns="http://www.w3.org/2000/svg" className="h-5 w-5 text-green-500 flex-shrink-0" viewBox="0 0 20 20" fill="currentColor">
-    <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.707-9.293a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z" clipRule="evenodd" />
-  </svg>
-);
-
-const XIcon = () => (
-  <svg xmlns="http://www.w3.org/2000/svg" className="h-5 w-5 text-red-500 flex-shrink-0" viewBox="0 0 20 20" fill="currentColor">
-    <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zM8.707 7.293a1 1 0 00-1.414 1.414L8.586 10l-1.293 1.293a1 1 0 101.414 1.414L10 11.414l1.293 1.293a1 1 0 001.414-1.414L11.414 10l1.293-1.293a1 1 0 00-1.414-1.414L10 8.586 8.707 7.293z" clipRule="evenodd" />
-  </svg>
-);
-
-const ClockIcon = () => (
-  <svg xmlns='http://www.w3.org/2000/svg' className='h-5 w-5 mr-1.5 text-gray-500 inline' fill='none' viewBox='0 0 24 24' stroke='currentColor'>
-    <path strokeLinecap='round' strokeLinejoin='round' strokeWidth={2} d='M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z'/>
-  </svg>
-);
-
-const LocationIcon = () => (
-  <svg xmlns='http://www.w3.org/2000/svg' className='h-5 w-5 mr-1.5 text-gray-500 inline' fill='none' viewBox='0 0 24 24' stroke='currentColor'>
-    <path strokeLinecap='round' strokeLinejoin='round' strokeWidth={2} d='M17.657 16.657L13.414 20.9a1.998 1.998 0 01-2.827 0l-4.244-4.243a8 8 0 1111.314 0z'/>
-    <path strokeLinecap='round' strokeLinejoin='round' strokeWidth={2} d='M15 11a3 3 0 11-6 0 3 3 0 016 0z'/>
-  </svg>
-);
-
-const CalendarIcon = () => (
-  <svg xmlns='http://www.w3.org/2000/svg' className='h-5 w-5 mr-1.5 text-gray-500 inline' fill='none' viewBox='0 0 24 24' stroke='currentColor'>
-    <path strokeLinecap='round' strokeLinejoin='round' strokeWidth={2} d='M8 7V3m8 4V3m-9 8h10M5 21h14a2 2 0 002-2V7a2 2 0 00-2-2H5a2 2 0 00-2 2v12a2 2 0 002 2z'/>
-  </svg>
-);
+import { CheckIcon, XIcon, CalendarIcon, ClockIcon, LocationIcon } from '@/assets/icons';
+import { letterToGradeValue } from '@/constants/gradeConstants';
 
 const Requirement = ({ text, met }) => (
   <li className={`flex items-center space-x-2 text-sm ${met ? 'text-gray-700' : 'text-red-600 font-medium'}`}>
@@ -43,11 +13,6 @@ const Requirement = ({ text, met }) => (
     <span>{text}</span>
   </li>
 );
-
-const letterToGradeValue = {
-  'A': 10, 'A-': 9, 'B+': 8, 'B': 7, 'B-': 6,
-  'C+': 5, 'C': 4, 'C-': 3, 'D': 2, 'F': 1,
-};
 
 export default function PositionsCard({ position, index }) {
   const { currentUser, refreshUserProfile } = useAuth();
@@ -180,7 +145,7 @@ export default function PositionsCard({ position, index }) {
       </div>
 
       {isFormOpen && (
-        <ApplicationForm
+        <EditableApplicationForm
           user={currentUser}
           position={position}
           onClose={() => setIsFormOpen(false)}
